requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
graphql-core==3.2.3
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
strawberry-graphql==0.155.0
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
